vertexai-mlops/pipeline-sample/requirements.txt
# Vertex AI Pipeline サンプル - 依存関係
# Requires Python 3.11+

# Kubeflow Pipelines SDK v2
kfp>=2.0.0,<3.0.0

# Google Cloud SDK
google-cloud-aiplatform>=1.60.0
google-cloud-pipeline-components>=2.0.0
google-cloud-bigquery>=3.25.0
google-auth>=2.29.0

# データ処理
pandas>=2.2.0
numpy>=1.26.0
pyarrow>=15.0.0
db-dtypes>=1.2.0

# ML ライブラリ
scikit-learn>=1.4.0

# 設定ファイル
pyyaml>=6.0.1

# ユーティリティ
python-dateutil>=2.9.0
